+++ src/addin/client-interfaces/addin-client-flyout-permalink.ts
@@ -1,5 +1,5 @@
 /**
- * Interface for information to displays a permalink button
+ * Interface for information to display a permalink button
  * in the flyout header that navigates users to the URL
  * representative of the flyout's contents.
  */
@@ -8,10 +8,10 @@ export interface AddinClientFlyoutPermalink {
   /**
    * Specifies a text label for the permalink button.
    */
-  label?: string;
+  label: string;
 
   /**
    * Specifies an external URL for the permalink.
    */
-  url?: string;
+  url: string;
 }